Support waiting before running a command
Sometimes we want to discuss the command before we run it. In particular
when running the command will hide the "typed" title and command.

Example command:

    $ # Watching drpc progress
    $ watch kubectl get drpc -A --context hub

With the new option we wait after typing the command until Enter is
pressed.

Like the `delay` option, `wait` can be applied at the demo level or only
for some steps.